status and user database integration
View Change:
When user is not signed in, statuses can be seen with time posted in
newsfeed. and clicking on the link, user is able to view individual
status.
"Create Status" link only works when user is signed in.
After signing in, user can view all statuses in newsfeed and able to
create status. In individual status view, if the status is updated by
the current user, it can be edited and deleted as well.

Route Change:
After updating status, user will be redirected back to news feed.