User rights added and Tagged view modified.
Now, users can delete statuses in which they are tagged.
Since creating tags could be done through User - first_name or last_name
or profile_name, if it is a "@"tag, all statuses related to that user
could be viewed.
Duplicate posts removed in user profile, for users having same first or
last or profile name.