Reporting feature added.
Reporting feature needs users to be logged in.
Confirmation display popped up upon clicking.
Even if one user reported it, it will be available on admin dashboard.
After reviewing, admin can dismiss reporting if found not offensive, or
delete the status or even monitor the status owner.

To consider:
Reporting will be always available, and keep track of the report count
on status.
Or maybe even keep track of the reported user as well.